slide-generate/cdk.out/asset.13f05396c2841a2dae93f31643eeb6858fc385bc622e0254810bc779fef2ea1e/ requirements.txt
python-pptx==0.6.21
boto3==1.28.44